+++ src/Query/Query.php
@@ -61,7 +61,6 @@ class Query implements Interface\Query
         $streamResult = new Results\Stream(
             $this->stream,
             $this->distinct,
-            $this->selectAll,
             $this->selectedFields,
             $this->excludedFields,
             $this->getFrom(),