+++ src/Query/TestProvider.php
@@ -36,10 +36,6 @@ class TestProvider implements Query
         return $this->selectedFields;
     }
 
-    public function getSelectAll(): bool
-    {
-        return $this->selectAll;
-    }
     /**
      * @return string[]
      */