+++ tests/Traits/SelectTest.php
@@ -81,24 +81,21 @@ class SelectTest extends TestCase
         $this->query->select('*');
 
         $selectedFields = $this->query->getSelectedFields();
-        $this->assertCount(0, $selectedFields);
-        $this->assertTrue($this->query->getSelectAll());
+        $this->assertCount(1, $selectedFields);
 
         $this->query = new TestProvider();
         $this->query->select('id, name, price')
             ->select('*');
 
         $selectedFields = $this->query->getSelectedFields();
-        $this->assertCount(3, $selectedFields);
-        $this->assertTrue($this->query->getSelectAll());
+        $this->assertCount(4, $selectedFields);
 
         $this->query = new TestProvider();
         $this->query->select('id, name, price')
             ->select(Query::SELECT_ALL);
 
         $selectedFields = $this->query->getSelectedFields();
-        $this->assertCount(3, $selectedFields);
-        $this->assertTrue($this->query->getSelectAll());
+        $this->assertCount(4, $selectedFields);
     }
 
     public function testSelectionSpecialFields(): void